Updated CSS, HTML and JS
CHANGES TO JS:
Needed to start from 1 for the phone since it there is an additional current_time_class for the current 12PM default in the HTML. For the web view, I will start at 0 but if i = 1 then continue. So need to fix the current time for the web.

MISC:
Also need to fix the layout for the phone after it slides in. Would like to incorporate the arrow first before continuing with the design. For now, I have implemented the foundation.

Have not touched the Tablet Version yet.